Lot of stuff added for ELK stack on minikube along with prometheus and grafana Changes to be committed: 	modified:   kubernetes/deploy/yaml/apache_phpadmin_mysql/combined/fwd.sh 	modified:   kubernetes/install_kubectl.sh 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/app-deployment.yml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/es-deployment.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/es-svc.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/esfwd.sh 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/filebeat.yml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/kibana-deployment.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/kibana-svc.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/kibfwd.sh 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/logstash-deployment.yml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/logstash-svc.yml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/ELK-Stack-Kubernetes-minikube/logstash.conf 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/elk-helm/elasticsearch-values.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/elk-helm/filebeat-values.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/elk-helm/kibana-values.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/elk-helm/logstash-values.yaml 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/prometheusandGrafana/getpasswd.sh 	new file:   kubernetes/logging/ELK-Stack-Kubernetes-minikube/prometheusandGrafana/install_promgrafana.sh 	modified:   kubernetes/start_minikube.sh

+++ kubernetes/install_kubectl.sh
@@ -1,7 +1,25 @@
-curl -LO https://dl.k8s.io/release/v1.31.5/bin/linux/amd64/kubectl
-curl -LO https://dl.k8s.io/release/v1.31.5/bin/linux/amd64/kubectl.sha256
+# !/bin/bash
+set -e 
+#  Install kubectl
+# https://kubernetes.io/docs/tasks/tools/install-kubectl-linux/
+# https://kubernetes.io/docs/tasks/tools/install-kubectl-linux/#install-using-native-package-management
+# https://kubernetes.io/docs/setup/release/version-skew-policy/#kubectl
+# https://kubernetes.io/docs/tasks/tools/#kubectl
+# The version of kubectl must be within one minor version of the control plane
+# see also install_minikube.sh and above links
+# update the command below to change the version of kubectl to install 
+# curl -LO "https://dl.k8s.io/release/$(curl -Ls https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
+
+curl -LO https://dl.k8s.io/release/$(curl -Ls https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl
+curl -LO https://dl.k8s.io/release/$(curl -Ls https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl.sha256
 echo "$(cat kubectl.sha256)  kubectl" | sha256sum --check
-sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
-rm -rf ./kubectl kubectl.sha256
-kubectl version --client
-kubectl version --client --output=yaml
+if [ $? -eq 0 ]; then
+    echo "kubectl checksum verification succeeded"
+    sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
+    rm -rf ./kubectl kubectl.sha256
+    kubectl version --client
+    kubectl version --client --output=yaml
+else
+    echo "kubectl checksum verification failed"
+    exit 1
+fi
